src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-added-display-none-host-expected.txt

CONSOLE ERROR: line 13: Uncaught TypeError: document.getElementById(...).createShadowRoot is not a function
Tests adding a shadow root to a display none host doesn't leave dirty bits.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
